Adds HS email to the list of emails
Ensures that HS email addresses are included when searching for a person.

This change checks if the contact item has an HS email and if it is not already in the list of emails. If both conditions are met, the HS email is added to the list, ensuring that all possible email addresses are considered during the search.